+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Settings (Invoices) | Apply baseline keyboard shortcuts. Refs UINV-303.
 * Tech debt: reducing code smells, usage of consts and resources from stripes-acq. Refs UINV-325.
+* Indicate that accounting code is a required field when export to accounting is true. Refs UINV-324.
 
 ## [3.0.2](https://github.com/folio-org/ui-invoice/tree/v3.0.2) (2021-11-25)
 [Full Changelog](https://github.com/folio-org/ui-invoice/compare/v3.0.1...v3.0.2)